Update config.h
Changed flashconfig.telnet_port to unsigned 16bit integer.

+++ esp-link/config.h
@@ -41,7 +41,7 @@ typedef struct {
   int8_t   data_bits;
   int8_t   parity;
   int8_t   stop_bits;
-  uint8_t  telnet_port1,               // Telnet port settings
+  uint16_t  telnet_port1,               // Telnet port settings
            telnet_port2;
 } FlashConfig;
 extern FlashConfig flashConfig;
